Move injector to a injectors package
